20170612　コメント見直し
