--- app/dto/ColorCloudAll.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-package dto;
-
-/**
- *
- * @author a-yamamoto
- */
-public class ColorCloudAll extends ColorChange {
-    public ColorCloudAll(int cloudAll){
-        
-        super.elem = cloudAll + "%";
-    } 
-}
