--- app/dto/ColorMain.java
@@ -1,17 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-package dto;
-
-/**
- *
- * @author a-yamamoto
- */
-public class ColorMain extends ColorChange {
-    public ColorMain(String main){
-        super.elem = main;
-    }
-}
