+++ app/dto/colorChange/ColorChange.java
@@ -1,0 +1,11 @@
+package dto.colorChange;
+
+/**
+ * セルに色を付けるためのスーパークラス：変数定義
+ * @author a-yamamoto
+ */
+public abstract class ColorChange {
+    public String color;
+    //各要素のvalue
+    public String elem;
+}
