+++ app/dto/colorChange/ColorMain.java
@@ -1,0 +1,11 @@
+package dto.colorChange;
+
+/**
+ *
+ * @author a-yamamoto
+ */
+public class ColorMain extends ColorChange {
+    public ColorMain(String main){
+        super.elem = main;
+    }
+}
