+++ app/dto/colorChange/ColorPressure.java
@@ -1,10 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-package dto;
+package dto.colorChange;
 
 /**
  *
