+++ app/dto/colorChange/ColorTemp.java
@@ -4,7 +4,7 @@
  * and open the template in the editor.
  */
 
-package dto;
+package dto.colorChange;
 
 /**
  *
